+++ crate/game_play_hud/src/constants.rs
@@ -1,0 +1,6 @@
+/// Length to render the `HpBar`.
+pub const HP_BAR_LENGTH: f32 = 100.;
+/// Height to render the `HpBar`.
+pub const HP_BAR_HEIGHT: f32 = 3.;
+/// Number of sprites on the `HpBar` sprite sheet.
+pub const HP_BAR_SPRITE_COUNT: usize = 10;