+++ crate/game_play_hud/src/system.rs
@@ -1,0 +1,3 @@
+pub use self::hp_bar_update_system::HpBarUpdateSystem;
+
+mod hp_bar_update_system;